requirements.txt
# =============================================================================
# VietAgent Python Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# =============================================================================

# --- Google APIs ---
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# --- Modal (Cloud Webhooks) ---
modal>=0.64.0

# --- Utilities ---
python-dotenv>=1.0.0
requests>=2.31.0
